% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_ragged_edge.R
\name{conditional_fill}
\alias{conditional_fill}
\title{Conditional Fill and Forecast for Time Series Data}
\usage{
conditional_fill(ts_data, sys_eq, dates, estimates, fill_method)
}
\arguments{
\item{ts_data}{Time-series data set for the estimation.}

\item{sys_eq}{A \code{koma_seq} object (\link{system_of_equations}) containing details
about the system of equations used in the model.}

\item{dates}{Key-value list for date ranges in various model operations.}

\item{estimates}{A \code{koma_estimate} object (see \code{\link{estimate}})
containing the estimates of the simultaneous equations model.}

\item{fill_method}{Character string indicating which central tendency measure
("mean" or "median") to use when filling ragged edges.}
}
\value{
An extended time series list containing the filled and forecasted
time series up to the current quarter.
}
\description{
This function fills missing observations and extends the time series data
up to and including the current quarter. The function conditionally forecasts
based on the estimates and realized observations.
}
\keyword{internal}
