% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_balanced_data.R
\name{construct_balanced_data}
\alias{construct_balanced_data}
\title{Construct Balanced Data}
\usage{
construct_balanced_data(
  ts_data,
  endogenous_variables,
  total_exogenous_variables,
  start,
  end,
  warn = FALSE
)
}
\arguments{
\item{ts_data}{Time series data.}

\item{endogenous_variables}{Endogenous variables.}

\item{total_exogenous_variables}{Exogenous and / or predetermined variables
to include in x matrix.}

\item{start}{Start date for truncation.}

\item{end}{End date for truncation.}

\item{warn}{Logical. If TRUE, warn when the start is moved because of
missing or lag-induced NAs. Default is FALSE.}
}
\value{
A list containing the truncated time series data, y_matrix, x_matrix
the number of observations, the date of last observation and the frequency.
}
\description{
Construct Balanced Data
}
\keyword{internal}
