% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_within_gibbs_algorithm.R
\name{construct_cholesky_of_inverse_hessian}
\alias{construct_cholesky_of_inverse_hessian}
\title{Cholesky factor of the inverse Hessian}
\usage{
construct_cholesky_of_inverse_hessian(hessian)
}
\arguments{
\item{hessian}{The Hessian of the target function at its optimum, as
returned by \code{\link[stats:optim]{stats::optim()}}.}
}
\value{
The lower triangular Cholesky factor of the inverse Hessian. Stops
with an error if the Hessian is not positive definite, i.e. the target has
no proper optimum to start the sampler from.
}
\description{
Computes the Cholesky factor \eqn{L} of the inverse of the Hessian
\eqn{M^{-1}} of the target function at its optimum, used to draw the
candidate gamma in the MH algorithm.
}
\keyword{internal}
