% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_construction.R
\name{construct_equation_data}
\alias{construct_equation_data}
\title{Cache fixed data for an equation}
\usage{
construct_equation_data(
  y_matrix,
  x_matrix,
  character_gamma_matrix,
  character_beta_matrix,
  jx
)
}
\arguments{
\item{y_matrix}{A \eqn{(T \times n)} matrix \eqn{Y}, where \eqn{T} is the
number of observations and \eqn{n} the number of equations, i.e. endogenous
variables.}

\item{x_matrix}{A \eqn{(T \times k)} matrix \eqn{X} of observations on
\eqn{k} exogenous variables.}

\item{character_gamma_matrix}{A matrix \eqn{\Gamma} that holds the
coefficients in character form for all equations. The dimensions of the
matrix are \eqn{(T \times n)}, where \eqn{T} is the number of
observations and \eqn{n} the number of equations.}

\item{character_beta_matrix}{A matrix \eqn{\beta} that holds the
coefficients in character form for all equations. The dimensions of the
matrix are \eqn{(k \times n)}, where \eqn{k} is the number of
exogenous variables and \eqn{n} the number of equations.}

\item{jx}{The index of equation \eqn{j}.}
}
\value{
A list containing the endogenous subset, restricted exogenous
subset, beta positions, parameter counts, and observation count.
}
\description{
Cache fixed data for an equation
}
\keyword{internal}
