% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ticks.R
\name{create_annotations}
\alias{create_annotations}
\title{Create Plotly Annotations at Ticks}
\usage{
create_annotations(
  x_ticks,
  y_position,
  font = list(family = NULL, size = NULL),
  digits = 1
)
}
\arguments{
\item{x_ticks}{A data frame containing the tick information. Must include
columns \code{year} for the x-axis positions, \code{value} for the text of the
annotations, and \code{color_code} for the color of the text.}

\item{y_position}{Numeric value specifying the y-axis position of all
annotations, typically in paper coordinates (0 to 1).}

\item{font}{A list specifying font attributes for the annotation text, with
optional \code{family} and \code{size} components to customize the font family and
size, respectively. If not specified, default Plotly text properties are
used.}

\item{digits}{Number of decimal places to round the annotation value to.
Defaults to 1.}
}
\value{
A list of lists, where each inner list represents an annotation in
the format required by Plotly. Each annotation includes properties such as
position, text, and formatting.
}
\description{
Generates a list of annotations for Plotly plots based on x tick position.
This function is designed to create text annotations for specific points on a
plot, such as in-sample and out-of-sample ticks, with customizable y
positions and formatting.
}
\keyword{internal}
