% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_within_gibbs_algorithm_informative.R
\name{draw_omega_j_informative}
\alias{draw_omega_j_informative}
\title{Draw Omega from inverse Wishart distribution for equation j}
\usage{
draw_omega_j_informative(
  y_matrix,
  x_matrix,
  character_gamma_matrix,
  character_beta_matrix,
  jx,
  gamma_jw,
  theta_jw,
  priors_j,
  equation_data
)
}
\arguments{
\item{y_matrix}{A \eqn{(T \times n)} matrix \eqn{Y}, where \eqn{T} is the
number of observations and \eqn{n} the number of equations, i.e. endogenous
variables.}

\item{x_matrix}{A \eqn{(T \times k)} matrix \eqn{X} of observations on
\eqn{k} exogenous variables.}

\item{character_gamma_matrix}{A matrix \eqn{\Gamma} that holds the
coefficients in character form for all equations. The dimensions of the
matrix are \eqn{(T \times n)}, where \eqn{T} is the number of
observations and \eqn{n} the number of equations.}

\item{character_beta_matrix}{A matrix \eqn{\beta} that holds the
coefficients in character form for all equations. The dimensions of the
matrix are \eqn{(k \times n)}, where \eqn{k} is the number of
exogenous variables and \eqn{n} the number of equations.}

\item{jx}{The index of equation \eqn{j}.}

\item{gamma_jw}{A \eqn{(n_j \times 1)} vector with the parameters
of the gamma matrix, where \eqn{n_j} is the number of endogenous variables in
equation \eqn{j}.}

\item{theta_jw}{A \eqn{(k \times nj+1)}, where \eqn{n_j} is the number of
endogenous variables in equation \eqn{j}.}

\item{priors_j}{The priors for \eqn{\omega} in equation \eqn{j}.}

\item{equation_data}{Fixed equation subsets and counts returned by
\code{\link[=construct_equation_data]{construct_equation_data()}}. The samplers compute this once per equation.}
}
\value{
List containing \eqn{{\tilde{\Omega}}_j^{(w)}} as \code{omega_tilde_jw}
and \eqn{\Omega_j^{(w)}} as \code{omega_jw}.
}
\description{
\code{draw_omega_j_informative} draws the variance-covariance matrix
\eqn{\tilde{\Omega}_j} for each row of \eqn{[ u_j,  V_j ]}.
}
\keyword{internal}
