% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_within_gibbs_algorithm_informative.R
\name{draw_parameters_j_informative}
\alias{draw_parameters_j_informative}
\title{Draw Parameters for equation j}
\usage{
draw_parameters_j_informative(
  y_matrix,
  x_matrix,
  character_gamma_matrix,
  character_beta_matrix,
  jx,
  gibbs_sampler,
  priors,
  progress = function(amount) invisible()
)
}
\arguments{
\item{y_matrix}{A \eqn{(T \times n)} matrix \eqn{Y}, where \eqn{T} is the
number of observations and \eqn{n} the number of equations, i.e. endogenous
variables.}

\item{x_matrix}{A \eqn{(T \times k)} matrix \eqn{X} of observations on
\eqn{k} exogenous variables.}

\item{character_gamma_matrix}{A matrix \eqn{\Gamma} that holds the
coefficients in character form for all equations. The dimensions of the
matrix are \eqn{(T \times n)}, where \eqn{T} is the number of
observations and \eqn{n} the number of equations.}

\item{character_beta_matrix}{A matrix \eqn{\beta} that holds the
coefficients in character form for all equations. The dimensions of the
matrix are \eqn{(k \times n)}, where \eqn{k} is the number of
exogenous variables and \eqn{n} the number of equations.}

\item{jx}{The index of equation \eqn{j}.}

\item{gibbs_sampler}{An object of class \code{gibbs_sampler} that holds an
equations gibbs settings.}

\item{priors}{The priors for \eqn{\theta} in equation \eqn{j}.}

\item{progress}{Function called with the number of completed draws since
its last call, at most about every 0.5 seconds. Used to update the progress
bar.}
}
\value{
A list containing matrices for the saved draws of parameters and
additional diagnostic information.
}
\description{
\code{draw_parameters_j_informative} simulates from the posterior of the model's
parameters for each equation \eqn{j} separately using an informative prior.
The posterior is simulated using a Metropolis-within-Gibbs sampling
procedure.
}
\details{
The sampler works as follows:
\enumerate{
\item Initialize sampler: \eqn{delta_{\gamma}^{(0)}} and \eqn{\Omega^{(0)}}
\item Conditional on \eqn{delta_{\gamma}^{(w-1)}} and \eqn{\Omega^{(w-1)}}
and the data draw \eqn{delta_{\theta}}.
\item Conditional on \eqn{delta_{\gamma}^{(w-1)}}, \eqn{delta_{\theta}^{(w)}}
draw \eqn{\Omega^{(w)}}.
\item Conditional on \eqn{delta_{\theta}^{(w)}}, \eqn{Omega^{(w)}} draw
\eqn{\delta_{\gamma}^{(w)}}.
\item Go back to step 2.
}
}
\keyword{internal}
