% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{update_identity_weights}
\alias{update_identity_weights}
\title{Update Identity Weights}
\usage{
update_identity_weights(weights, identities)
}
\arguments{
\item{weights}{A named list where each entry represents an equation (e.g.,
"gdp"). Each equation is itself a named list, containing variable names
(e.g., "manufacturing") and their dynamic weight values.

Example:
\code{list(gdp = list(manufacturing = c(0.2, 0.3), service = c(0.7, 0.6)))}}

\item{identities}{A named list, where each entry represents an equation
similar to \code{weights}. The equation has sublists "components" for variable
names and "weights" for identity weights.

Example:
\code{list(gdp = list(components = list(manufacturing = "theta_gamma6_4"),
                weights = list(theta_gamma6_4 = NULL)))}}
}
\value{
A named list containing the updated identity weights for each
equation and its variables, in the same structure as the input \code{identities}.
}
\description{
Iteratively updates the weights in an identity list (\code{identities}) based on
the last value of the corresponding weights found in the \code{weights} list.
}
\keyword{internal}
