% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{validate_date_range}
\alias{validate_date_range}
\title{Validate a Start/End Date Range in \code{dates}}
\usage{
validate_date_range(dates, field, frequency = 4, call = rlang::caller_env())
}
\arguments{
\item{dates}{A list of date ranges, e.g. \code{dates$estimation}.}

\item{field}{Name of the range in \code{dates} to validate, e.g. "estimation".}

\item{frequency}{Frequency of the time series.}

\item{call}{The environment from which the error is called.}
}
\value{
Invisibly \code{NULL}; aborts if the range is invalid.
}
\description{
Validate a Start/End Date Range in \code{dates}
}
\keyword{internal}
