% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{validate_model_evaluation_input}
\alias{validate_model_evaluation_input}
\title{Validate Model Evaluation Input}
\usage{
validate_model_evaluation_input(
  sys_eq,
  variables,
  horizon,
  ts_data,
  dates,
  evaluate_on_levels,
  call = rlang::caller_env()
)
}
\arguments{
\item{sys_eq}{A \code{koma_seq} object (\link{system_of_equations}) containing details
about the system of equations used in the model.}

\item{variables}{A character vector of name(s) of the stochastic
endogenous variable for which the forecast error(s) should be calculated. If
NULL it is calculated for all variables.}

\item{horizon}{The forecast horizon in quarters up to which the RMSE should
be calculated.}

\item{ts_data}{time series data set, must include data until end date of
forecasting period.}

\item{dates}{Key-value list for date ranges in various model operations.}

\item{evaluate_on_levels}{Boolean, if TRUE RMSE is calculated on levels if
FALSE on growth rates.}

\item{call}{The environment from which the error is called.}
}
\value{
The variables to evaluate: \code{variables}, or all endogenous variables
if \code{variables} is \code{NULL}.
}
\description{
Validate Model Evaluation Input
}
\keyword{internal}
