% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_sys_eq_priors}
\alias{validate_sys_eq_priors}
\title{Validate the Priors Stored in a System of Equations}
\usage{
validate_sys_eq_priors(sys_eq, call = rlang::caller_env())
}
\arguments{
\item{sys_eq}{A \code{koma_seq} object.}

\item{call}{The environment from which the error is called.}
}
\value{
\code{TRUE} invisibly, or an error listing the invalid priors.
}
\description{
\code{validate_priors()} checks the prior syntax in the equation strings when the
system is created. The priors can be changed afterwards in \code{sys_eq$priors},
so this checks the stored priors again before they are used.
}
\keyword{internal}
